+++ CHANGELOG.md
@@ -5,29 +5,51 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.3] - 2026-09-23
+
+### Added
+
+- Implemented `hybridSearch` functionality using BM25 sparse + dense vector
+- Created utility script to build filter expression for categories, authors, and date
+- Added function to retrieve existing ids to avoid re-embedding the records that already exist in the database.
+- Option to get results by `query` (scalar) if only search filters are provided
+
+### Changed
+
+- Modified `ArxivSchema` to include bm25 function, sparse vector, and revised indexes
+- Adapted the tests to work with the latest changes
+- Changed `serverDest` sync in `logger.ts` to use `true` to resolve complications with `milvus-utils.ts`
+- Revised logging calls in `ingestion-service.ts` to be brief, and more meaningful
+- Updated `MILVUS_ALIAS` and `MODEL_NAME` in .env.example
+
 ## [0.3.2] - 2026-09-17
 
 ### Fixed
+
 - Updated `.env.example` with separate Pino server and ingestion log path variables
 
 ## [0.3.1] - 2026-09-17
 
 ### Added
+
 - Installed `pino-http` and `pino-pretty` packages
 - Created `run-server.sh` to start up Express API server from root dir
 - Added `pino-http` middleware to Express to log HTTP requests
 
 ### Changed
+
 - Revamped Pino logging infrastructure (https://github.com/vkumar825/arxiv-search/pull/32)
 - Switched console logging for error and info to use server logging instance
 - Made `ingestion-service.ts` output logging information to its dedicated log file (`ingestion.log`)
 
 ### Fixed
+
 - Resolved broken tests in `search-controller.test.ts` by adding mocked `req.log`
 
 ## [0.3.0] - 2026-09-15
 
 ### Added
+
 - Implemented data cleaning pipeline using Python (https://github.com/vkumar825/arxiv-search/pull/28)
 - Refactored codebase to move away from schema-blind approach and stick to arXiv dataset
 - Overhauled arXiv schema to account for the cleaned arXiv dataset (https://github.com/vkumar825/arxiv-search/pull/27)
@@ -35,33 +57,39 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added unit tests for search controllers, milvus schemas, and embedding services.
 
 ### Changed
+
 - Updated dependencies, removed unused libraries in package.json
 - Addressed security vulnerabilities for `adm-zip`, `sharp`, and `qs` packages
 
 ## [0.2.2] - 2026-08-29
 
 ### Added
+
 - Integrated Vitest testing framework (https://github.com/vkumar825/arxiv-search/issues/15)
 - Added unit tests for search controllers, milvus schemas, and embedding services.
 
 ## [0.2.1] - 2026-08-02
 
 ### Changed
+
 - Updated dependencies to address security vulnerabilities (https://github.com/vkumar825/arxiv-search/pull/24)
 
 ## [0.2.0] - 2026-07-31
 
 ### Changed
+
 - Migrated to `typescript` (https://github.com/vkumar825/arxiv-search/pull/22)
 - Changed `README.md` to reflect the latest changes
 - Included a step to use the Express.js Search API in the `README.md`
 
 ## Added
+
 - Set up basic configuration for TypeScript
 
 ## [0.1.0] - 2026-07-19
 
 ### Added
+
 - Initial project scaffolding and basic setup.
 - Basic functionalities of the Milvus vector database implemented in the first release.
 - Included detailed documentation and setup instructions in the `README.md`